pyproject.toml
[project]
name = "mdp-dedupe"
version = "0.1.0"
description = "Medical data provider deduplication tool for identifying and merging duplicate patient records"
authors = [{ name = "itskitto", email = "kitto.khr@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "faker>=33.3.1,<34.0.0",
    "python-dateutil>=2.9.0.post0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "six>=1.17.0,<2.0.0",
    "sqlalchemy>=2.0.37,<3.0.0",
    "typing-extensions>=4.12.2,<5.0.0",
    "pandas>=2.2.3,<3.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "dedupe>=3.0.3,<4.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "tabulate (>=0.9.0,<0.10.0)",
    "pyarrow (>=19.0.0,<20.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "mypy>=1.8.0,<2.0.0",
    "black>=24.1.0,<25.0.0",
    "isort>=5.13.2,<6.0.0",
    "flake8>=7.0.0,<8.0.0",
    "pylint>=3.0.3,<4.0.0",
]

[project.scripts]
mdp-dedupe = "mdp_dedupe.__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=mdp_dedupe --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["dedupe.*"]
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ["py313"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
